@@ -2,10 +2,11 @@
 
 module github.com/crimist/trakx
 
-go 1.19
+go 1.20
 
 require (
 	github.com/cbeuw/connutil v0.0.0-20200411215123-966bfaa51ee3
+	github.com/davecgh/go-spew v1.1.1
 	github.com/go-torrent/bencode v0.0.0-20150403200907-4318e12a955c
 	github.com/heroku/x v0.0.55
 	github.com/kkyr/fig v0.3.0
