@@ -1,21 +1,15 @@
 package config
 
 import (
-	"os"
-	"strconv"
-	"strings"
-	"sync"
+	"path/filepath"
 	"time"
 
-	"github.com/pkg/errors"
 	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
 )
 
 type Configuration struct {
-	loaded bool // config is loaded and valid
-
 	LogLevel       LogLevel
+	CachePath      string
 	ExpvarInterval time.Duration
 	Debug          struct {
 		Pprof int
@@ -60,23 +54,48 @@ type Configuration struct {
 		Trim   time.Duration
 		Expiry time.Duration
 	}
-	Path struct {
-		Log string
-		Pid string
-	}
 }
 
-// Loaded returns true if the config was successfully parsed and loaded.
-func (config *Configuration) Loaded() bool { return config.loaded }
+type Warnings int
 
-// SetLogLevel sets the desired loglevel in the in memory configuration and logger
-func (conf *Configuration) SetLogLevel(level LogLevel) {
+const (
+	WarningNone          Warnings = iota
+	WarningUDPValidation Warnings = 1 << iota
+	WarningPeerExpiry    Warnings = 1 << iota
+)
+
+// Validate ensures that configuration values are sane and returns warnings for potential misconfigurations or security issues
+func (conf *Configuration) Validate() Warnings {
+	warnings := WarningNone
+
+	if !conf.UDP.ConnDB.Validate {
+		warnings |= WarningUDPValidation
+	}
+	if conf.DB.Expiry < conf.Announce.Base+conf.Announce.Fuzz {
+		warnings |= WarningPeerExpiry
+	}
+
+	return warnings
+}
+
+// LogPath returns the log path as defined by the configuration and current time
+func (conf *Configuration) LogPath() string {
+	return filepath.Join(conf.CachePath, "trakx_"+time.Now().Format("06-01-02-15-04-05")+".log")
+}
+
+// PIDPath retuirns the pid file path
+func (conf *Configuration) PIDPath() string {
+	return filepath.Join(conf.CachePath, "trakx.pid")
+}
+
+// setLogLevel sets the desired loglevel in the in memory configuration and logger
+func (conf *Configuration) setLogLevel(level LogLevel) {
 	conf.LogLevel = level
 
 	switch level {
 	case "debug":
 		loggerAtom.SetLevel(zap.DebugLevel)
-		Logger.Debug("Debug loglevel set, debug panics enabled")
+		zap.L().Debug("Debug loglevel set, debug panics enabled")
 	case "info":
 		loggerAtom.SetLevel(zap.InfoLevel)
 	case "warn":
@@ -86,61 +105,7 @@ func (conf *Configuration) SetLogLevel(level LogLevel) {
 	case "fatal":
 		loggerAtom.SetLevel(zap.FatalLevel)
 	default:
-		Logger.Warn("Invalid log level was specified, defaulting to warn")
+		zap.L().Warn("Invalid log level was specified, defaulting to warn")
 		loggerAtom.SetLevel(zap.WarnLevel)
 	}
-}
-
-var oneTimeSetup sync.Once
-
-// Parse updates logger and limits based on the configuration settings.
-func (config *Configuration) Parse() error {
-	// one time logger atom setup
-	oneTimeSetup.Do(func() {
-		loggerAtom = zap.NewAtomicLevelAt(zap.DebugLevel)
-	})
-
-	cfg := zap.NewDevelopmentConfig()
-
-	// set strings to lowercase
-	config.LogLevel = LogLevel(strings.ToLower(string(config.LogLevel)))
-	config.HTTP.Mode = strings.ToLower(config.HTTP.Mode)
-
-	// dev env check
-	if config.LogLevel.Debug() {
-		cfg.Development = true
-	} else {
-		cfg.Development = false
-	}
-
-	// setup logger
-	Logger = zap.New(zapcore.NewCore(zapcore.NewConsoleEncoder(cfg.EncoderConfig), zapcore.Lock(os.Stdout), loggerAtom))
-	config.SetLogLevel(config.LogLevel)
-
-	// resolve env vars for database backup path
-	if strings.HasPrefix(config.DB.Backup.Path, "ENV:") {
-		config.DB.Backup.Path = os.Getenv(strings.TrimPrefix(config.DB.Backup.Path, "ENV:"))
-	}
-
-	// resolve paths
-	home, err := os.UserHomeDir()
-	if err != nil {
-		Logger.Fatal("failed to get home directory", zap.Error(err))
-	}
-	config.Path.Pid = strings.ReplaceAll(config.Path.Pid, "~", home)
-	config.Path.Log = strings.ReplaceAll(config.Path.Log, "~", home)
-
-	// If $PORT var set override port for appengines (like heroku)
-	if appenginePort := os.Getenv("PORT"); appenginePort != "" {
-		appPort, err := strconv.Atoi(appenginePort)
-		if err != nil {
-			return errors.Wrap(err, "failed to parse $PORT env variable (not an int)")
-		}
-
-		Logger.Info("PORT env variable detected. Overriding config...", zap.Int("$PORT", appPort))
-		config.HTTP.Port = appPort
-	}
-
-	config.loaded = true
-	return nil
 }
