@@ -2,16 +2,15 @@ package pools
 
 import (
 	"github.com/crimist/trakx/bencoding"
-	"github.com/crimist/trakx/tracker/storage"
 )
 
 var (
 	Peerlists4   *Pool[[]byte]
 	Peerlists6   *Pool[[]byte]
 	Dictionaries *Pool[*bencoding.Dictionary]
-	Peers        *Pool[*storage.Peer]
 )
 
+// TODO: create these pools in their respective packages, most of these don't need to be global variables
 func Initialize(numwantLimit int) {
 	peerlist4Max := 6 * numwantLimit  // ipv4 + port
 	peerlist6Max := 18 * numwantLimit // ipv6 + port
@@ -37,8 +36,4 @@ func Initialize(numwantLimit int) {
 	}, func(dictionary *bencoding.Dictionary) {
 		dictionary.Reset()
 	})
-
-	Peers = NewPool[*storage.Peer](func() any {
-		return new(storage.Peer)
-	}, nil)
 }
