@@ -3,7 +3,6 @@ package http
 import (
 	"net"
 
-	"github.com/crimist/trakx/config"
 	"github.com/crimist/trakx/pools"
 	"github.com/crimist/trakx/tracker/stats"
 	"go.uber.org/zap"
@@ -18,13 +17,13 @@ func writeErr(conn net.Conn, msg string) {
 	pools.Dictionaries.Put(dictionary)
 }
 
-func (t *HTTPTracker) clientError(conn net.Conn, msg string) {
+func (t *Tracker) clientError(conn net.Conn, msg string) {
 	stats.ClientErrors.Add(1)
 	writeErr(conn, msg)
 }
 
-func (t *HTTPTracker) internalError(conn net.Conn, errmsg string, err error) {
+func (t *Tracker) internalError(conn net.Conn, errmsg string, err error) {
 	stats.ServerErrors.Add(1)
 	writeErr(conn, "internal server error")
-	config.Logger.Error(errmsg, zap.Error(err))
+	zap.L().Error(errmsg, zap.Error(err))
 }
