@@ -3,15 +3,15 @@ package http
 import (
 	"net"
 
-	"github.com/crimist/trakx/tracker/utils/unsafemanip"
+	"github.com/crimist/trakx/utils"
 )
 
 func redir(c net.Conn, url string) {
-	c.Write(unsafemanip.StringToBytes("HTTP/1.1 303\r\nLocation: " + url + "\r\n\r\n"))
+	c.Write(utils.StringToBytesUnsafe("HTTP/1.1 303\r\nLocation: " + url + "\r\n\r\n"))
 }
 
 func writeData(c net.Conn, data string) {
-	c.Write(unsafemanip.StringToBytes("HTTP/1.1 200\r\n\r\n" + data))
+	c.Write(utils.StringToBytesUnsafe("HTTP/1.1 200\r\n\r\n" + data))
 }
 
 func writeDataBytes(c net.Conn, data []byte) {
@@ -19,5 +19,5 @@ func writeDataBytes(c net.Conn, data []byte) {
 }
 
 func writeStatus(c net.Conn, status string) {
-	c.Write(unsafemanip.StringToBytes("HTTP/1.1 " + status + "\r\n\r\n"))
+	c.Write(utils.StringToBytesUnsafe("HTTP/1.1 " + status + "\r\n\r\n"))
 }
