@@ -1,54 +1,29 @@
 package udp
 
 import (
-	"github.com/crimist/trakx/config"
-	"github.com/crimist/trakx/tracker/stats"
-	"github.com/crimist/trakx/tracker/udp/protocol"
+	"net"
+
+	"github.com/crimist/trakx/tracker/udp/udpprotocol"
 	"go.uber.org/zap"
 )
 
-type cerrFields map[string]interface{}
-
-func (u *UDPTracker) newClientError(msg string, TransactionID int32, fieldMap ...cerrFields) []byte {
-	stats.ClientErrors.Add(1)
-
-	if config.Config.LogLevel.Debug() {
-		fields := []zap.Field{zap.String("msg", msg)}
-		if len(fieldMap) == 1 {
-			for k, v := range fieldMap[0] {
-				fields = append(fields, zap.Any(k, v))
-			}
-		}
-
-		config.Logger.Info("Client Err", fields...)
+func (tracker *Tracker) fatal(remote *net.UDPAddr, message []byte, TransactionID int32) {
+	if tracker.stats != nil {
+		// TODO: this isn't right
+		tracker.stats.ServerErrors.Add(1)
 	}
 
-	e := protocol.Error{
-		Action:        protocol.ActionError,
+	protoError := udpprotocol.ErrorResponse{
+		Action:        udpprotocol.ActionError,
 		TransactionID: TransactionID,
-		ErrorString:   []byte(msg),
+		ErrorString:   message,
 	}
 
-	data, err := e.Marshall()
+	data, err := protoError.Marshal()
 	if err != nil {
-		config.Logger.Error("e.Marshall()", zap.Error(err))
+		zap.L().Error("failed to marshal error packet", zap.Error(err))
+		tracker.socket.WriteToUDP([]byte("catastrophic failure"), remote)
+	} else {
+		tracker.socket.WriteToUDP(data, remote)
 	}
-	return data
-}
-
-func (u *UDPTracker) newServerError(msg string, err error, TransactionID int32) []byte {
-	stats.ServerErrors.Add(1)
-
-	e := protocol.Error{
-		Action:        protocol.ActionError,
-		TransactionID: TransactionID,
-		ErrorString:   []byte("internal err"),
-	}
-	config.Logger.Error(msg, zap.Error(err))
-
-	data, err := e.Marshall()
-	if err != nil {
-		config.Logger.Error("e.Marshall()", zap.Error(err))
-	}
-	return data
 }
